Screen can no longer go out of the map

+++ src/jixel/entity/JixelEntity.java
@@ -11,7 +11,7 @@ import jixel.stage.JixelGame;
 public abstract class JixelEntity extends JixelSprite implements Serializable {
 
 	private String name;
-	private int x, y;
+	private double x, y;
 	private double speed;
 
 	public JixelEntity(final String PATH, String name, int tileX, int tileY, double speed) {
@@ -60,28 +60,28 @@ public abstract class JixelEntity extends JixelSprite implements Serializable {
 	/**
 	 * @return the x
 	 */
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
 	/**
 	 * @param x the x to set
 	 */
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
 	/**
 	 * @return the y
 	 */
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
 	/**
 	 * @param y the y to set
 	 */
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
